backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.1.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
gunicorn>=21.0.0
werkzeug>=3.0.0
redis>=5.0.0
requests>=2.31.0
flask-migrate>=4.0.0
